FWUV-58 - IdP Protocol Mapper with mapped values - Review issues fixed
